+++ .gitignore
@@ -1,4 +1,6 @@
 .DS_Store
+.idea/
+.vscode/
 build/juno
 coverage/*
 courtney/
