+++ pkg/rpc/rpc_test.go
@@ -1,4 +1,4 @@
-package test
+package rpc
 
 import (
 	"bytes"
@@ -11,12 +11,10 @@ import (
 	"os"
 	"testing"
 	"time"
-
-	"github.com/NethermindEth/juno/pkg/rpc"
 )
 
-func getServerHandler() *rpc.HandlerJsonRpc {
-	return rpc.NewHandlerJsonRpc(rpc.HandlerRPC{})
+func getServerHandler() *HandlerJsonRpc {
+	return NewHandlerJsonRpc(HandlerRPC{})
 }
 
 type rpcTest struct {
@@ -72,7 +70,7 @@ func TestRPCServer(t *testing.T) {
 }
 
 func TestServer(t *testing.T) {
-	server := rpc.NewServer(":8080")
+	server := NewServer(":8080")
 	go func() {
 		_ = server.ListenAndServe()
 	}()
